requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.3.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
Werkzeug==3.0.2
